Refactoring, Connect tflite
