+++ .idea/gradle.xml
@@ -10,6 +10,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/ailia-sdk-jni" />
+            <option value="$PROJECT_DIR$/ailia-tflite-jni" />
             <option value="$PROJECT_DIR$/ailia-tokenizer-jni" />
             <option value="$PROJECT_DIR$/app" />
           </set>
