+++ .idea/vcs.xml
@@ -4,6 +4,7 @@
     <mapping directory="$PROJECT_DIR$/.." vcs="Git" />
     <mapping directory="$PROJECT_DIR$" vcs="Git" />
     <mapping directory="$PROJECT_DIR$/ailia-sdk-jni" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/ailia-tflite-jni" vcs="Git" />
     <mapping directory="$PROJECT_DIR$/ailia-tokenizer-jni" vcs="Git" />
   </component>
 </project>
