+++ app/build.gradle
@@ -40,6 +40,7 @@ dependencies {
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
     implementation project(':ailia-sdk-jni')
     implementation project(':ailia-tokenizer-jni')
+    implementation project(':ailia-tflite-jni')
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.ext:junit:1.2.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
