+++ app/src/main/java/jp/axinc/ailia_kotlin/AiliaTFLiteSample.kt
@@ -1,0 +1,4 @@
+package jp.axinc.ailia_kotlin
+
+class AiliaTFLiteSample {
+}
