+++ settings.gradle
@@ -10,3 +10,4 @@ rootProject.name = "ailia-kotlin"
 include ':app'
 include ':ailia-sdk-jni'
 include ':ailia-tokenizer-jni'
+include ':ailia-tflite-jni'
